+++ src/main/java/com/kafkamgt/clusterapi/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.kafkamgt.clusterapi;
+package com.kafkamgt.clusterapi.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
