+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-DEV.schemaregistry.url=http://localhost:8081
+#DEV.schemaregistry.url=http://localhost:8081
 
 server.port:9343
 
@@ -10,6 +10,6 @@ securityconfig.user2.username=user2
 securityconfig.user2.pwd=pwd
 securityconfig.user2.role=ADMIN
 
-license.key=VCFG-5RT6-OKI9-QW3F
-license.spec=My ABC Company
-license.specsalt=My ABC Company, NewYork
+#license.key=VCFG-5RT6-OKI9-QW3F
+#license.spec=My ABC Company
+#license.specsalt=My ABC Company, NewYork
